External Java library
