+++ .gitignore
@@ -8,4 +8,3 @@
 /build
 /captures
 .externalNativeBuild
-/libs
