+++ app/src/main/java/c/example/speechtobsl/Parser.java
@@ -1,0 +1,11 @@
+package c.example.speechtobsl;
+
+import c.example.parser.StanfordParser;
+
+public class Parser extends StanfordParser {
+
+    public Parser() {
+        super();
+    }
+
+}
