+++ parser/build.gradle
@@ -1,0 +1,13 @@
+apply plugin: 'java-library'
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation "edu.stanford.nlp:stanford-corenlp:3.9.2"
+    implementation "edu.stanford.nlp:stanford-corenlp:3.9.2:models"
+    implementation "org.slf4j:slf4j-api:1.7.25"
+    implementation "org.slf4j:slf4j-log4j12:1.7.25"
+    
+}
+
+sourceCompatibility = "8"
+targetCompatibility = "8"
